Update streaming, tracing, services and model to reflect latest java SDK
