+++ src/zepben/cimbend/common/translator/base_proto2cim.py
@@ -8,7 +8,7 @@ from typing import Optional
 
 from dataclassy import dataclass
 
-from zepben.cimbend._dataclass import DataClassMetaZ
+
 from zepben.cimbend.common.base_service import BaseService
 from zepben.protobuf.cim.iec61970.base.core.IdentifiedObject_pb2 import IdentifiedObject as PBIdentifiedObject
 from zepben.protobuf.cim.iec61968.common.Document_pb2 import Document as PBDocument
@@ -59,7 +59,7 @@ PBOrganisationRole.to_cim = organisationrole_to_cim
 PBIdentifiedObject.to_cim = identifiedobject_to_cim
 
 
-@dataclass(slots=True, meta=DataClassMetaZ)
+@dataclass(slots=True)
 class BaseProtoToCim(object, metaclass=ABCMeta):
     service: BaseService
