+++ src/zepben/cimbend/tracing/phase_status.py
@@ -3,6 +3,7 @@
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
 from __future__ import annotations
 
 from zepben.cimbend.cim.iec61970.base.wires import SinglePhaseKind
