+++ src/zepben/cimbend/tracing/phasing.py
@@ -11,7 +11,7 @@ import logging
 from dataclassy import dataclass
 from enum import Enum
 
-from zepben.cimbend._dataclass import DataClassMetaZ
+
 from zepben.cimbend.cim.iec61970.base.wires.energy_source import EnergySource
 from zepben.cimbend.cim.iec61970.base.wires.switch import Breaker
 from zepben.cimbend.cim.iec61970.base.wires.single_phase_kind import SinglePhaseKind
@@ -20,7 +20,7 @@ from zepben.cimbend.exceptions import PhaseException
 from zepben.cimbend.tracing.connectivity import get_connectivity
 from zepben.cimbend.exceptions import TracingException
 from zepben.cimbend.tracing.phase_status import normal_phases, current_phases
-from zepben.cimbend.tracing.util import queue_next_terminal
+from zepben.cimbend.tracing.traces import queue_next_terminal
 from zepben.cimbend.traversals.queue import PriorityQueue
 from zepben.cimbend.traversals.tracing import Traversal
 from zepben.cimbend.tracing.phase_status import PhaseStatus
@@ -40,7 +40,7 @@ class FeederProcessingStatus(Enum):
     NONE = 2
 
 
-@dataclass(slots=True, meta=DataClassMetaZ)
+@dataclass(slots=True)
 class FeederCbTerminalPhasesByStatus:
     terminal: Terminal
     in_phases: Set[SinglePhaseKind] = set()
@@ -48,7 +48,7 @@ class FeederCbTerminalPhasesByStatus:
     phases_to_flow: Set[SinglePhaseKind] = set()
 
 
-@dataclass(slots=True, meta=DataClassMetaZ)
+@dataclass(slots=True)
 class DelayedFeederTrace:
     out_terminal: Terminal
     phases_to_flow: Set[SinglePhaseKind]
